requirements.txt
Flask==0.10.1
flask-restless==0.8
Jinja2==2.7.2
MarkupSafe==0.18
Werkzeug==0.9.4
argparse==1.2.1
distribute==0.6.24
itsdangerous==0.23
psycopg2==2.7.2
uWSGI==2.0.2
wsgiref==0.1.2
Flask-SQLAlchemy==1.0
